+++ crates/bevy_scene/src/dynamic_scene.rs
@@ -193,7 +193,7 @@ where
 #[cfg(test)]
 mod tests {
     use bevy_ecs::entity::EntityHashMap;
-    use bevy_ecs::{reflect::AppTypeRegistry, system::Command, world::World};
+    use bevy_ecs::{reflect::AppTypeRegistry, world::Command, world::World};
     use bevy_hierarchy::{Parent, PushChild};
 
     use crate::dynamic_scene_builder::DynamicSceneBuilder;
